+++ examples/asgi/main.py
@@ -3,7 +3,7 @@ import logging
 import uvicorn
 
 from engin import Supply
-from engin.ext.asgi import ASGIEngin
+from engin.extensions.asgi import ASGIEngin
 from examples.asgi.app import AppBlock, AppConfig
 from examples.asgi.common.db.block import DatabaseBlock
 from examples.asgi.features.cats.block import CatBlock