+++ CHANGELOG.md
@@ -6,6 +6,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.0.19] - 2025-04-27
+
+### Added
+
+- A new exception: `InvalidBlockError`.
+
+### Changed
+
+- Renamed the `ext` subpackage to `extensions`.
+- Errors are now imported from `engin.exceptions.*` instead of `engin.*`
+- Blocks will now raise an `InvalidBlockError` if the block has methods which are not
+  decorated with `@provide` & `@invoke`.
+
+
 ## [0.0.18] - 2025-04-25
 
 ### Added